requirements.txt
discord.py>=2.5.2
pytz>=2025.2
python-dotenv>=1.1.1
psutil>=6.1.1
google-auth-oauthlib>=1.2.2
google-api-python-client>=2.176.0
google>=3.0.0
protobuf>=6.31.1
matplotlib>=3.10.0
pillow>=11.3.0
aiohttp>=3.12.13
ipywidgets>=8.1.5
asyncpg>=0.30.0
gspread>=6.0.2
Flask>=3.1.1